fix(g36): disconnect SC receive error paths
